fix(windows): fix std::max type mismatch and move CUDA files to correct source group
Two remaining Windows MSVC errors after unistd.h and OpenMP fixes:

1. std::max(jj, i) type mismatch in knn_sparse.inl: After changing loop index i from size_t to std::ptrdiff_t for OpenMP MSVC compatibility, std::max(jj, i) failed because jj is size_t and i is now ptrdiff_t. Fixed with static_cast<size_t>(i).

2. cuda_edge_extraction.hpp in non-CUDA build: The source file was incorrectly placed in NERVE_CORE_SOURCES (compiled always) instead of NERVE_CUDA_CORE_SOURCES (CUDA-only). Also proactively moved 5 other GPU-dependent files from the same directory:
   - fast_vr_gpu_integration.cpp
   - gpu_memory_monitor.cpp
   - gpu_reduction_engine.cpp
   - heterogeneous_fast_vr.cpp
   - hybrid_reduction_engine.cpp
   These contain CUDA header includes and kernel declarations that MSVC cannot compile.

+++ src/algorithms/detail/distance_matrix_knn_sparse.inl
@@ -24,7 +24,7 @@ std::vector<T> DistanceMatrixComputer<T>::compute_blocked(std::span<const T> poi
 #endif
             for (std::ptrdiff_t i = static_cast<std::ptrdiff_t>(ii); i < static_cast<std::ptrdiff_t>(i_end); ++i)
             {
-                for (size_t j = std::max(jj, i); j < j_end; ++j)
+                for (size_t j = std::max(jj, static_cast<size_t>(i)); j < j_end; ++j)
                 {
                     T sum = 0;
                     for (size_t d = 0; d < dim; ++d)
